third_party/blink/web_tests/external/wpt/css/motion/offset-supports-calc-expected.txt

This is a testharness.js-based test.
FAIL offset-position supports calc assert_equals: expected "calc(20% + 30px) calc(100% - 40px)" but got "calc(20% + 30px) calc(100% + -40px)"
PASS offset-path supports calc
FAIL offset-distance supports calc assert_equals: expected "calc(50% - 100px)" but got "calc(50% + -100px)"
PASS offset-rotate supports calc
FAIL offset-anchor supports calc assert_equals: expected "calc(20% + 30px) calc(100% - 40px)" but got "calc(20% + 30px) calc(100% + -40px)"
Harness: the test ran to completion.
